requirements.txt
nltk
numpy==1.23.0
pandas
python-dotenv
scikit-learn
sentence-transformers
setuptools
transformers
wheel
requests
beautifulsoup4==4.12.2
scipy
httpx==0.27.0
tenacity==8.2.3
ujson==5.9.0
urllib3==2.2.0
xyzservices==2023.10.1
pyyaml==6.0.1
regex==2023.12.25
six==1.16.0
tqdm==4.66.1
typing-extensions==4.10.0
typing-inspect==0.9.0